Stimmkreis components now take wahlJahr prop (#28)
Note: this is a little buggy at the moment
